+++ packages/oauth2/lib/src/exception.dart
@@ -5,9 +5,12 @@ class AuthorizationException extends AngelHttpException {
   final ErrorResponse errorResponse;
 
   AuthorizationException(this.errorResponse,
-      {StackTrace? stackTrace, int? statusCode, error})
-      : super(error ?? errorResponse,
-            stackTrace: stackTrace, message: '', statusCode: statusCode ?? 400);
+      {StackTrace? stackTrace, int? statusCode, dynamic error})
+      : super(
+            stackTrace: stackTrace,
+            message: '',
+            //error: error,
+            statusCode: statusCode ?? 400);
 
   @override
   Map toJson() {